requirements.txt
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
Django==3.2.4